+++ src/main/java/core/model/player/ISkillChange.java
@@ -29,7 +29,7 @@ public interface ISkillChange {
 	*
 	* @return value
 	*/
-	public abstract int getValue();
+	public abstract double getValue();
 
 	public abstract int getChange();
 }
